fixed two compilation errors

+++ src/xo-print.c
@@ -824,7 +824,7 @@ gboolean pdf_draw_image(PdfImage *image, struct XrefTable *xref, GString *pdfbuf
     "%d 0 obj\n<< /Length %d /Filter /FlateDecode /Type /Xobject "
     "/Subtype /Image /Width %d /Height %d /ColorSpace /DeviceRGB "
     "/BitsPerComponent 8 ",
-    image->n_obj, zpix->len, width, height);
+    image->n_obj, (int)zpix->len, width, height);
   if (image->has_alpha) {
     g_string_append_printf(pdfbuf, 
       "/SMask %d 0 R ",
@@ -853,7 +853,7 @@ gboolean pdf_draw_image(PdfImage *image, struct XrefTable *xref, GString *pdfbuf
       "%d 0 obj\n<< /Length %d /Filter /FlateDecode /Type /Xobject "
       "/Subtype /Image /Width %d /Height %d /ColorSpace /DeviceGray "
       "/BitsPerComponent 8 >> stream\n",
-      image->n_obj_smask, zpix->len, width, height);
+      image->n_obj_smask, (int)zpix->len, width, height);
 
     g_string_append_len(pdfbuf, zpix->str, zpix->len);
     g_string_free(zpix, TRUE);
